pyproject.toml
[project]
name = "langchain-basic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bilibili-api-python>=17.4.1",
    "bs4>=0.0.2",
    "deepagents>=0.3.0",
    "faiss-cpu>=1.13.1",
    "flagembedding>=1.3.5",
    "langchain>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.0",
    "langchain-qwq>=0.3.1",
    "langchain-tavily>=0.2.13",
    "langchain-text-splitters>=1.0.0",
    "langchainhub>=0.1.21",
    "langgraph-checkpoint-postgres>=3.0.1",
    "langgraph-cli[inmem]>=0.4.9",
    "lark>=1.3.1",
    "llama-index>=0.14.10",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-embeddings-ollama>=0.8.4",
    "llama-index-experimental>=0.6.3",
    "llama-index-llms-dashscope>=0.5.1",
    "llama-index-llms-langchain>=0.7.1",
    "llama-index-llms-ollama>=0.9.0",
    "llama-index-llms-openai>=0.6.12",
    "llama-index-vector-stores-milvus>=0.9.4",
    "markdown>=3.10",
    "openpyxl>=3.1.5",
    "peft>=0.18.0",
    "pymilvus[model]>=2.6.5",
    "pypdf>=6.4.0",
    "sentence-transformers>=5.2.0",
    "spacy>=3.8.11",
    "tavily-python>=0.7.14",
    "unstructured[pdf]>=0.18.21",
]
